+++ src/main/java/ladder/control/input/single/NameInput.java
@@ -5,7 +5,7 @@ import ladder.model.participant.Name;
 import ladder.view.View;
 
 public class NameInput extends InputControl<Name> {
-    public static Name ALL = new Name("all");
+    public static final Name ALL = new Name("all");
 
     public NameInput(View view) {
         super(view);
