+++ src/main/java/ladder/control/output/LadderGameOutput.java
@@ -3,20 +3,20 @@ package ladder.control.output;
 
 import ladder.model.LadderGame;
 import ladder.view.View;
-import ladder.view.result.output.LadderGameView;
+import ladder.view.result.game.LadderGameView;
 
 public class LadderGameOutput {
-    private final View outputView;
-    private final LadderGameView gameView;
+    private final View info;
+    private final LadderGameView output;
 
     public LadderGameOutput(View outputView, LadderGame game) {
-        this.outputView = outputView;
-        this.gameView = new LadderGameView(game);
+        this.info = outputView;
+        this.output = new LadderGameView(game);
 
     }
 
     public void print(int formatWidth) {
-        outputView.render();
-        gameView.render(formatWidth);
+        info.render();
+        output.render(formatWidth);
     }
 }
