+++ src/main/java/ladder/view/result/game/NamesView.java
@@ -1,4 +1,4 @@
-package ladder.view.result.output;
+package ladder.view.result.game;
 
 import ladder.model.participant.Names;
