+++ src/main/java/ladder/view/result/match/info/LadderMatchInfoView.java
@@ -1,8 +1,8 @@
-package ladder.view.result;
+package ladder.view.result.match.info;
 
 import ladder.view.View;
 
-public class LadderResultInfoView implements View {
+public class LadderMatchInfoView implements View {
     @Override
     public void render() {
         System.out.println("\n실행 결과");
